Save file working properly - needs to be tested on big save files
