+++ src/@types/Document.ts
@@ -1,3 +1,5 @@
+import { ScoringCriteria } from '.';
+
 export interface Document {
     riesitel: string;
     kategoria: string;
@@ -11,6 +13,20 @@ export interface Document {
     tags: any[];
     originalName: string;
     opened: boolean;
+}
+
+export interface DocumentBackup {
+    changes: any[];
+    scoring?: IScoring;
+    tags: Tag[];
+    opened: boolean;
+    timeOpened: number;
+}
+
+export interface BackupFile {
+    changes: Record<string, DocumentBackup>;
+    tags: Tag[];
+    scoringCriteria: ScoringCriteria[];
 }
 
 export interface IScoring {
