+++ src/@types/index.ts
@@ -1,3 +1,3 @@
-export { Document, IScoring, DocumentMetadata, DocumentParser, Tag, ITemplate } from './Document';
+export { Document, IScoring, DocumentMetadata, DocumentParser, Tag, ITemplate, DocumentBackup, BackupFile } from './Document';
 export { Tool, ToolOptions, ScoringCriteria } from './Tools';
 export { OthersCategory, SettingsCategory, ShortcutCategory, ToolsCategory, Settings, Shortcut } from './Preferences';
