+++ src/components/Topbar/AddDocumentsButton.vue
@@ -1,20 +1,19 @@
 <template>
   <b-nav-item v-b-modal.noveRiesenie>
-    <span
-    >Pridať ďalšie riešenie</span>
+    <span>Pridať ďalšie riešenie</span>
     <b-modal
-        id="noveRiesenie"
-        size="lg"
-        title="Pridat nové riešenie"
-        @ok="pridajRiesenia"
-        :ok-disabled="noveRiesenia.length == 0"
+      id="noveRiesenie"
+      size="lg"
+      title="Pridat nové riešenie"
+      @ok="pridajRiesenia"
+      :ok-disabled="noveRiesenia.length == 0"
     >
       <b-form-file
-          accept=".pdf"
-          multiple
-          id="rieseniaInput"
-          v-model="noveRiesenia"
-          :file-name-formatter="formatNames"
+        accept=".pdf"
+        multiple
+        id="rieseniaInput"
+        v-model="noveRiesenia"
+        :file-name-formatter="formatNames"
       />
     </b-modal>
   </b-nav-item>
@@ -22,7 +21,7 @@
 
 <script lang="ts">
 import { Database } from '@/Db';
-import { activeParser, AddDocument, Documents } from '@/DocumentManager';
+import { activeParser, AddDocument, Documents } from '@/Documents/DocumentManager';
 import Vue from 'vue'
 import Component from 'vue-class-component';
 
@@ -40,23 +39,23 @@ export default class AddDocumentButton extends Vue {
     for (let i = 0; i < this.noveRiesenia.length; i++) {
       const riesenie = this.noveRiesenia[i];
       addOperations.push(
-          new Promise<void>((resolve, reject) => {
-            const index = i;
-            riesenie.arrayBuffer().then((buffer) => {
-              AddDocument(riesenie.name, buffer, baseLength + index + 1)
-                  .then(() => resolve())
-                  .catch((err) => this.$bvModal.msgBoxConfirm(`Riešenie ${riesenie.name} už existuje, chceš ho nahradiť novým?`, { okTitle: 'Áno', cancelTitle: 'Nie' }).then(val => {
-                    if (val) {
-                      const id = activeParser.parse(riesenie.name).id
-                      Database.removeDocument(id).then(() => {
-                        AddDocument(riesenie.name, buffer, baseLength + index + 1)
-                            .then(() => resolve())
-                      })
-                    }
-                    else resolve()
-                  }));
-            });
-          })
+        new Promise<void>((resolve, reject) => {
+          const index = i;
+          riesenie.arrayBuffer().then((buffer) => {
+            AddDocument(riesenie.name, buffer, baseLength + index + 1)
+              .then(() => resolve())
+              .catch((err) => this.$bvModal.msgBoxConfirm(`Riešenie ${riesenie.name} už existuje, chceš ho nahradiť novým?`, { okTitle: 'Áno', cancelTitle: 'Nie' }).then(val => {
+                if (val) {
+                  const id = activeParser.parse(riesenie.name).id
+                  Database.removeDocument(id).then(() => {
+                    AddDocument(riesenie.name, buffer, baseLength + index + 1)
+                      .then(() => resolve())
+                  })
+                }
+                else resolve()
+              }));
+          });
+        })
       );
     }
     Promise.all(addOperations).then(() => {
