+++ src/views/Editor.vue
@@ -46,7 +46,7 @@ import Viewport from "../components/Viewport.vue";
 import Topbar from "../components/Topbar/Topbar.vue";
 import Toolbar from "../components/Tools/Toolbar.vue";
 import SearchBar from "../components/SearchBar.vue";
-import { Documents, getViewedDocument, loadFromDatabase } from "../DocumentManager";
+import { Documents, getViewedDocument, loadFromDatabase } from "../Documents/DocumentManager";
 import type { PDFdocument } from "@/components/PDFdocument";
 import Scoring from "@/components/Scoring.vue";
 import { loadFonts } from "@/components/Fonts";
